moves county methods into subfolder, but keeps within FipsLookup module
